+++ src/background/modules/system_settings/mod.rs
@@ -1,3 +1,3 @@
-mod application;
+pub mod application;
 mod domain;
 pub mod infrastructure;