+++ src/background/seelen_bar/mod.rs
@@ -153,24 +153,23 @@ impl FancyToolbar {
     /// Work area no works fine on multiple monitors
     /// so we use this functions that only takes the toolbar in account
     pub fn get_work_area_by_monitor(monitor: HMONITOR) -> Result<RECT> {
-        let monitor_info = WindowsApi::monitor_info(monitor)?;
-        let dpi = WindowsApi::get_device_pixel_ratio(monitor)?;
-        let mut rect = monitor_info.monitorInfo.rcMonitor;
         let state = FULL_STATE.load();
         let settings = state.settings().fancy_toolbar();
+
+        let monitor_info = WindowsApi::monitor_info(monitor)?;
+        let monitor_scale_factor = WindowsApi::get_monitor_scale_factor(monitor)?;
+        let text_scale_factor = WindowsApi::get_text_scale_factor()?;
+        let real_height =
+            (settings.height as f64 * monitor_scale_factor * text_scale_factor) as i32;
+
+        let mut rect = monitor_info.monitorInfo.rcMonitor;
         if state.is_bar_enabled() {
             match settings.position {
                 FancyToolbarSide::Top => {
-                    rect = RECT {
-                        top: rect.top + (settings.height as f32 * dpi) as i32,
-                        ..rect
-                    };
+                    rect.top += real_height;
                 }
                 FancyToolbarSide::Bottom => {
-                    rect = RECT {
-                        bottom: rect.bottom - (settings.height as f32 * dpi) as i32,
-                        ..rect
-                    };
+                    rect.bottom -= real_height;
                 }
             }
         }
@@ -185,21 +184,19 @@ impl FancyToolbar {
         let settings = &state.settings().fancy_toolbar();
 
         let monitor_info = WindowsApi::monitor_info(monitor)?;
-        let monitor_dpi = WindowsApi::get_device_pixel_ratio(monitor)?;
+        let monitor_scale_factor = WindowsApi::get_monitor_scale_factor(monitor)?;
+        let text_scale_factor = WindowsApi::get_text_scale_factor()?;
         let rc_monitor = monitor_info.monitorInfo.rcMonitor;
 
+        let real_height =
+            (settings.height as f64 * monitor_scale_factor * text_scale_factor) as i32;
+        self.theoretical_rect = rc_monitor;
         match settings.position {
             FancyToolbarSide::Top => {
-                self.theoretical_rect = RECT {
-                    bottom: rc_monitor.top + (settings.height as f32 * monitor_dpi) as i32,
-                    ..rc_monitor
-                };
+                self.theoretical_rect.bottom = rc_monitor.top + real_height;
             }
             FancyToolbarSide::Bottom => {
-                self.theoretical_rect = RECT {
-                    top: rc_monitor.bottom - (settings.height as f32 * monitor_dpi) as i32,
-                    ..rc_monitor
-                };
+                self.theoretical_rect.top = rc_monitor.bottom - real_height;
             }
         }