+++ changelog.md
@@ -12,6 +12,7 @@
 - panic on media module when loading initial devices.
 - discord window without umid (for now umid was hardcoded).
 - focused app not updating on title change.
+- not considerating accesibility text scale factor on toolbar and dock.
 
 ## [2.2.2]
 ### enhancements